i remember why i added this as well
okay so the reason we need to do this in a desktop app rather than a web app is because everytime a user makes a change, the code would have to update, meaning we need a server. but spending money is a nono especially since none of the pages we're coding need a server, so it would just be a waste of money. using a desktop app means no server and we can store data on the computer

and the reason we're doing this in godot is because i like it more than other desktop app development softwares and it's easy to compile for both macOS and windows